Add scippnexus.typing
